src/main/java/fr/b4/apps/clients/web/UserController.java
package fr.b4.apps.clients.web;

import fr.b4.apps.clients.entities.User;
import fr.b4.apps.clients.repositories.UserRepository;
import io.swagger.annotations.*;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@Api(value = "Users API", tags = "Users", protocols = "http")
@RequestMapping("/users")
@RestController
public class UserController {

    private final UserRepository userRepository;

    public UserController(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @GetMapping
    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    @PostMapping
    public User save(@RequestBody User toSave) {
        return userRepository.save(toSave);
    }
}
